Fix key code mapping utilities and tap-hold precedence
