+++ Sources/KeyPathAppKit/UI/Overlay/BaseKeycap+Labels.swift
@@ -10,7 +10,7 @@ extension BaseKeycap {
         if !isPressed,
            let tapHoldIdleLabel,
            shouldShowTapHoldIdleLabel,
-           !TapHoldLabelPrecedence.idleLabelMatchesBase(tapHoldIdleLabel, baseLabel: baseLabel, keyLabel: key.label)
+           !TapHoldLabelPrecedence.idleLabelAddsNoNewVisualInformation(tapHoldIdleLabel, baseLabel: baseLabel, keyLabel: key.label)
         {
             return tapHoldIdleLabel
         }
