+++ Sources/KeyPathAppKit/UI/Overlay/OverlayKeyboardView.swift
@@ -588,120 +588,7 @@ struct OverlayKeyboardView: View {
     /// Maps CGEvent key codes to OsCode Display names (lowercase).
     /// These match what Kanata sends via TCP KeyInput events: OsCode.to_string().to_lowercase()
     nonisolated static func keyCodeToKanataName(_ keyCode: UInt16) -> String {
-        switch keyCode {
-        // Row 3: Home row (ASDF...)
-        case 0: "a"
-        case 1: "s"
-        case 2: "d"
-        case 3: "f"
-        case 4: "h"
-        case 5: "g"
-        // Row 4: Bottom row (ZXCV...)
-        case 6: "z"
-        case 7: "x"
-        case 8: "c"
-        case 9: "v"
-        case 11: "b"
-        // Row 2: Top row (QWERTY...)
-        case 12: "q"
-        case 13: "w"
-        case 14: "e"
-        case 15: "r"
-        case 16: "y"
-        case 17: "t"
-        // Row 1: Number row
-        case 18: "1"
-        case 19: "2"
-        case 20: "3"
-        case 21: "4"
-        case 22: "6"
-        case 23: "5"
-        case 24: "equal"
-        case 25: "9"
-        case 26: "7"
-        case 27: "minus"
-        case 28: "8"
-        case 29: "0"
-        // More top row keys
-        case 30: "rightbrace"
-        case 31: "o"
-        case 32: "u"
-        case 33: "leftbrace"
-        case 34: "i"
-        case 35: "p"
-        // Home row continued
-        case 36: "enter"
-        case 37: "l"
-        case 38: "j"
-        case 39: "apostrophe"
-        case 40: "k"
-        case 41: "semicolon"
-        case 42: "backslash"
-        // Bottom row continued
-        case 43: "comma"
-        case 44: "slash"
-        case 45: "n"
-        case 46: "m"
-        case 47: "dot"
-        // Special keys
-        case 48: "tab"
-        case 49: "space"
-        case 50: "grave"
-        case 51: "backspace"
-        case 53: "esc"
-        // Modifiers
-        case 54: "rightmeta"
-        case 55: "leftmeta"
-        case 56: "leftshift"
-        case 57: "capslock"
-        case 58: "leftalt"
-        case 59: "leftctrl"
-        case 60: "rightshift"
-        case 61: "rightalt"
-        case 63: "fn"
-        // Function keys
-        case 96: "f5"
-        case 97: "f6"
-        case 98: "f7"
-        case 99: "f3"
-        case 100: "f8"
-        case 101: "f9"
-        case 103: "f11"
-        case 109: "f10"
-        case 111: "f12"
-        case 118: "f4"
-        case 120: "f2"
-        case 122: "f1"
-        // Arrow keys
-        case 123: "left"
-        case 124: "right"
-        case 125: "down"
-        case 126: "up"
-        // ISO key (between Left Shift and Z on ISO keyboards)
-        case 10: "intlbackslash"
-        // ABNT2 extra key (between slash and right shift on Brazilian keyboards)
-        case 94: "intlro" // International Ro key (ABNT2)
-        // Korean language toggle keys
-        case 104: "hangeul" // Hanja key
-        // Navigation keys
-        case 115: "home"
-        case 116: "pageup"
-        case 117: "del"
-        case 119: "end"
-        case 121: "pagedown"
-        case 114: "help"
-        // Extended function keys
-        case 64: "f17"
-        case 79: "f18"
-        case 80: "f19"
-        case 102: "rightctrl" // Also used for Han/Eng toggle on Korean keyboards
-        case 105: "f13"
-        case 106: "f16"
-        case 107: "f14"
-        case 113: "f15"
-        default:
-            "unknown-\(keyCode)"
-        }
+        KanataKeyCodeMap.overlayName(for: keyCode)
     }
 }
